WIP: Create awsvpn script
This uses the modified openvpn as well as the small go server
in order to connect to the AWS vpn. It follows the basic idea from
the aws-vpn-client repository, but makes it scriptable.

Notably, it adds a awsvpn.conf file so the values can be saved
It also gets the saml token via stdout instead of a file